requirements.txt
fastapi==0.68.1
requests==2.26.0
Pillow==8.3.2